Refactor: Fix Clean Architecture dependency violation
I've moved HttpStatusCode from the Application to the Domain layer.

- I created `src/domain/shared/http.status.ts` and moved the `HttpStatusCode` enum to this file.
- I updated imports in `src/domain/entities/user.ts`, `src/application/use-cases/auth.use-case.ts`, and `src/infraestructure/database/postgres/postgres.user.repository.ts`.
- The `src/application/shared/` directory is now empty. I was unable to remove it due to sandbox limitations.

Note: I could not verify the build and tests due to issues with `npm install` in the sandbox environment.
